# Conflicts:
#	src/main/java/repositories/RepositoryUser.java

+++ src/main/java/repositories/RepositoryUser.java
@@ -1,9 +1,12 @@
 package repositories;
 
-import entities.User;
+import com.example.demo.Entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepositoryUser extends JpaRepository<User, Integer> {
+public interface RepositoryUser extends JpaRepository<User, Long> {
+
+    //uuooooolaaa
+    //hola2
 }
